requirements.txt
anthropic>=0.39.0
typer>=0.12.0
python-dotenv>=1.0.0
